+++ tests/test_graph.py
@@ -9,8 +9,15 @@ from typing import Set, Tuple
 import networkx as nx
 
 from y0.dsl import A, B, C, D, M, Variable, X, Y, Z
-from y0.examples import SARS_SMALL_GRAPH, Example, examples, verma_1
-from y0.graph import DEFAULT_TAG, DEFULT_PREFIX, NxMixedGraph
+from y0.examples import SARS_SMALL_GRAPH, Example, examples, napkin, verma_1
+from y0.graph import (
+    DEFAULT_TAG,
+    DEFULT_PREFIX,
+    NxMixedGraph,
+    is_a_fixable,
+    is_markov_blanket_shielded,
+    is_p_fixable,
+)
 from y0.resources import VIRAL_PATHOGENESIS_PATH
 
 
@@ -288,3 +295,340 @@ class TestGraph(unittest.TestCase):
         )
         g2_y0_pre = set(map(lambda node: node.name, g2_y0_pre))
         self.assertEqual(g2_y0_pre, g2_ananke_pre)
+
+
+class TestFixability(unittest.TestCase):
+    """A test case for fixability in estimation workflows and tools."""
+
+    def assert_mb_shielded(self, graph: NxMixedGraph):
+        """Assert the graph is mb-shielded."""
+        self.assertTrue(graph.to_admg().mb_shielded())
+        self.assertTrue(is_markov_blanket_shielded(graph))
+
+    def assert_mb_unshielded(self, graph: NxMixedGraph):
+        """Assert the graph is not mb-shielded."""
+        self.assertFalse(graph.to_admg().mb_shielded())
+        self.assertFalse(is_markov_blanket_shielded(graph))
+
+    def test_is_mb_shielded(self):
+        """Test checking for markov blanket shielding."""
+        # Adapted from https://gitlab.com/causal/ananke/-/blob/dev/tests/estimation/test_automated_if.py#L80-92
+        graph_unshielded = NxMixedGraph.from_str_edges(
+            directed=[("T", "M"), ("M", "L"), ("L", "Y")], undirected=[("M", "Y")]
+        )
+        with self.subTest(name="Graph 1"):
+            self.assert_mb_unshielded(graph_unshielded)
+
+        # Second test: Napkin model
+        with self.subTest(name="Napkin"):
+            self.assert_mb_unshielded(napkin)
+
+        # Third test
+        with self.subTest(name="Graph 3"):
+            self.assert_mb_unshielded(SARS_SMALL_GRAPH)
+
+        # Fourth test
+        graph_4 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("SOS", "Ras"),
+                ("Ras", "PI3K"),
+                ("Ras", "Raf"),
+                ("PI3K", "AKT"),
+                ("AKT", "Raf"),
+            ],
+            undirected=[("SOS", "PI3K")],
+        )
+        with self.subTest(name="Graph 4"):
+            self.assert_mb_shielded(graph_4)
+
+        # Fifth test
+        graph_5 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("Z1", "X"),
+                ("X", "M1"),
+                ("M1", "Y"),
+                ("Z1", "Z2"),
+                ("Z2", "Z3"),
+                ("Z3", "Y"),
+                ("Z2", "M1"),
+            ],
+            undirected=[("Z1", "X"), ("Z2", "M1")],
+        )
+        with self.subTest(name="Graph 5"):
+            self.assert_mb_unshielded(graph_5)
+
+        # Sixth test
+        graph_6 = NxMixedGraph.from_str_edges(
+            directed=[("Z1", "X"), ("X", "M1"), ("M1", "Y"), ("Z1", "Y")]
+        )
+        with self.subTest(name="Graph 6"):
+            self.assert_mb_shielded(graph_6)
+
+        # Seventh test
+        graph_7 = NxMixedGraph.from_str_edges(directed=[("A", "B"), ("B", "C")])
+        with self.subTest(name="Graph 6"):
+            self.assert_mb_shielded(graph_7)
+
+    def assert_a_fixable(self, graph: NxMixedGraph, treatment: Variable):
+        """Assert that the graph is a-fixable."""
+        self.assertTrue(_ananke_a_fixable(graph, treatment))
+        self.assertTrue(is_a_fixable(graph, treatment))
+
+    def assert_not_a_fixable(self, graph: NxMixedGraph, treatment: Variable):
+        """Assert that the graph is not a-fixable."""
+        self.assertFalse(_ananke_a_fixable(graph, treatment))
+        self.assertFalse(is_a_fixable(graph, treatment))
+
+    def test_is_a_fixable(self):
+        """Test checking for a-fixability.
+
+        Graphs 9, 10, and 11 are from https://gitlab.com/causal/ananke/-/blob/dev/tests/\
+        estimation/test_counterfactual_mean.py?ref_type=heads#L20-47
+        """
+        graph_1 = NxMixedGraph.from_str_edges(
+            directed=[("T", "M"), ("M", "L"), ("L", "Y")], undirected=[("M", "Y")]
+        )
+        treatment_1 = Variable("T")
+        self.assert_a_fixable(graph_1, treatment_1)
+
+        self.assert_not_a_fixable(napkin, X)
+
+        graph_3 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("ADAM17", "EGFR"),
+                ("ADAM17", "TNF"),
+                ("ADAM17", "Sil6r"),
+                ("EGFR", "cytok"),
+                ("TNF", "cytok"),
+                ("Sil6r", "IL6STAT3"),
+                ("IL6STAT3", "cytok"),
+            ],
+            undirected=[
+                ("ADAM17", "cytok"),
+                ("ADAM17", "Sil6r"),
+                ("EGFR", "TNF"),
+                ("EGFR", "IL6STAT3"),
+            ],
+        )
+        treatment_3 = Variable("EGFR")
+        self.assert_a_fixable(graph_3, treatment_3)
+
+        graph_4 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("SOS", "Ras"),
+                ("Ras", "PI3K"),
+                ("Ras", "Raf"),
+                ("PI3K", "AKT"),
+                ("AKT", "Raf"),
+            ],
+            undirected=[("SOS", "PI3K")],
+        )
+        treatment_4 = Variable("PI3K")
+        self.assert_a_fixable(graph_4, treatment_4)
+
+        graph_5 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("Z1", "X"),
+                ("X", "M1"),
+                ("M1", "Y"),
+                ("Z1", "Z2"),
+                ("Z2", "Z3"),
+                ("Z3", "Y"),
+                ("Z2", "M1"),
+            ],
+            undirected=[("Z1", "X"), ("Z2", "M1")],
+        )
+        self.assert_a_fixable(graph_5, X)
+
+        graph_6 = NxMixedGraph.from_str_edges(
+            directed=[("Z1", "X"), ("X", "M1"), ("M1", "Y"), ("Z1", "Y")], undirected=[]
+        )
+        self.assert_a_fixable(graph_6, X)
+
+        graph_7 = NxMixedGraph.from_str_edges(directed=[("A", "B"), ("B", "C")], undirected=[])
+        treatment_7 = Variable("A")
+        self.assert_a_fixable(graph_7, treatment_7)
+
+        graph_8 = graph_4
+        treatment_8 = Variable("SOS")
+        self.assert_not_a_fixable(graph_8, treatment_8)
+
+    def assert_p_fixable(self, graph: NxMixedGraph, treatment: Variable):
+        """Assert that the graph is p-fixable."""
+        self.assertTrue(_ananke_p_fixable(graph, treatment))
+        self.assertTrue(is_p_fixable(graph, treatment))
+
+    def assert_not_p_fixable(self, graph: NxMixedGraph, treatment: Variable):
+        """Assert that the graph is not p-fixable."""
+        self.assertFalse(_ananke_p_fixable(graph, treatment))
+        self.assertFalse(is_p_fixable(graph, treatment))
+
+    def test_is_p_fixable(self):
+        """Test checking for p-fixability.
+
+        .. seealso:: https://gitlab.com/causal/ananke/-/blob/dev/tests/\
+        estimation/test_counterfactual_mean.py?ref_type=heads#L151-212
+        """
+        graph_1 = NxMixedGraph.from_str_edges(
+            directed=[("T", "M"), ("M", "L"), ("L", "Y")], undirected=[("M", "Y")]
+        )
+        treatment_1 = Variable("T")
+        with self.subTest(name="Graph 1"):
+            self.assert_p_fixable(graph_1, treatment_1)
+
+        with self.subTest(name="Napkin"):
+            self.assert_not_p_fixable(napkin, X)
+
+        treatment_3 = Variable("EGFR")
+        with self.subTest(name="Graph 3"):
+            self.assert_p_fixable(SARS_SMALL_GRAPH, treatment_3)
+
+        graph_4 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("SOS", "Ras"),
+                ("Ras", "PI3K"),
+                ("Ras", "Raf"),
+                ("PI3K", "AKT"),
+                ("AKT", "Raf"),
+            ],
+            undirected=[("SOS", "PI3K")],
+        )
+        treatment_4 = Variable("PI3K")
+        with self.subTest(name="Graph 4"):
+            self.assert_p_fixable(graph_4, treatment_4)
+
+        graph_5 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("Z1", "X"),
+                ("X", "M1"),
+                ("M1", "Y"),
+                ("Z1", "Z2"),
+                ("Z2", "Z3"),
+                ("Z3", "Y"),
+                ("Z2", "M1"),
+            ],
+            undirected=[("Z1", "X"), ("Z2", "M1")],
+        )
+        with self.subTest(name="Graph 5"):
+            self.assert_p_fixable(graph_5, X)
+
+        graph_6 = NxMixedGraph.from_str_edges(
+            directed=[("Z1", "X"), ("X", "M1"), ("M1", "Y"), ("Z1", "Y")], undirected=[]
+        )
+        with self.subTest(name="Graph 6"):
+            self.assert_p_fixable(graph_6, X)
+
+        graph_7 = NxMixedGraph.from_str_edges(directed=[("A", "B"), ("B", "C")], undirected=[])
+        treatment_7 = Variable("A")
+        with self.subTest(name="Graph 7"):
+            self.assert_p_fixable(graph_7, treatment_7)
+
+        graph_8 = graph_4
+        treatment_8 = Variable("SOS")
+        with self.subTest(name="Graph 8"):
+            self.assert_p_fixable(graph_8, treatment_8)
+
+        graph_9 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("C", "T"),
+                ("C", "M"),
+                ("C", "L"),
+                ("C", "Y"),
+                ("T", "M"),
+                ("M", "L"),
+                ("M", "Y"),
+                ("L", "Y"),
+            ],
+            undirected=[("T", "L"), ("T", "Y")],
+        )
+        treatment_9 = Variable("T")
+        with self.subTest(name="Graph 9"):
+            self.assert_p_fixable(graph_9, treatment_9)
+
+        graph_10 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("C", "T"),
+                ("C", "M"),
+                ("C", "L"),
+                ("C", "Y"),
+                ("T", "M"),
+                ("M", "L"),
+                ("T", "Y"),
+                ("L", "Y"),
+            ],
+            undirected=[("T", "L"), ("M", "Y")],
+        )
+        treatment_10 = Variable("T")
+        with self.subTest(name="Graph 10"):
+            self.assert_p_fixable(graph_10, treatment_10)
+
+        graph_11 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("C1", "T"),
+                ("C1", "L"),
+                ("C2", "M"),
+                ("C2", "L"),
+                ("C2", "Y"),
+                ("T", "M"),
+                ("M", "L"),
+                ("L", "Y"),
+            ],
+            undirected=[("Z1", "C1"), ("Z2", "C2"), ("T", "L")],
+        )
+        treatment_11 = Variable("T")
+        with self.subTest(name="Graph 11"):
+            self.assert_p_fixable(graph_11, treatment_11)
+
+        graph_12 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("C2", "T"),
+                ("C2", "M1"),
+                ("C2", "M2"),
+                ("C2", "Y"),
+                ("C1", "T"),
+                ("C1", "M2"),
+                ("T", "M1"),
+                ("M1", "M2"),
+                ("M2", "Y"),
+                ("T", "Y"),
+            ],
+            undirected=[("T", "M2"), ("M1", "Y")],
+        )
+        treatment_12 = Variable("T")
+        with self.subTest(name="Graph 12"):
+            self.assert_p_fixable(graph_12, treatment_12)
+
+        graph_13 = NxMixedGraph.from_str_edges(
+            directed=[
+                ("ViralLoad", "Income"),
+                ("ViralLoad", "T"),
+                ("ViralLoad", "Toxicity"),
+                ("Education", "Income"),
+                ("Education", "T"),
+                ("Education", "Toxicity"),
+                ("Income", "Insurance"),
+                ("Insurance", "T"),
+                ("T", "Toxicity"),
+                ("Toxicity", "CD4"),
+                ("T", "CD4"),
+            ],
+            undirected=[
+                ("Income", "T"),
+                ("Insurance", "ViralLoad"),
+                ("Education", "CD4"),
+                ("Income", "CD4"),
+            ],
+        )
+        treatment_13 = Variable("T")
+        with self.subTest(name="Graph 13"):
+            self.assert_not_p_fixable(graph_13, treatment_13)
+
+
+def _ananke_a_fixable(graph: NxMixedGraph, treatment: Variable) -> bool:
+    admg = graph.to_admg()
+    return 1 == len(admg.district(treatment.name).intersection(admg.descendants([treatment.name])))
+
+
+def _ananke_p_fixable(graph: NxMixedGraph, treatment: Variable) -> bool:
+    admg = graph.to_admg()
+    return 0 == len(admg.district(treatment.name).intersection(admg.children([treatment.name])))
